Fix mount / path issues with storage_nfs_el7 provisioning script
Clearly was not tested well (because of the host installation, but
that's an execuse).

This patch fixes it, tested on 3.6 with adding all domains, finally.
It also again simplifies it - but removes the option of not mounting the export
mount. Note that the ISO mount is not completely done as well, I think.
Will attend to in a follow up patch.

Based on David's request, the whole setup is split into functions.
Tested (properly) with 3.6, including the TEMPLATES domain attachment.
(The test for it should still be excluded, IMHO).

Change-Id: I261f3d3f3b3647cbbc506913478ffc6c36594519
Signed-off-by: Yaniv Kaul <ykaul@redhat.com>