ovirt-system-tests: test ovirt-log collector
This patch adds a test for ovirt-log-collector.

The collection is done in parallel to secondary storage
domains creation, so takes little or no time at all.
ovirt-log-collector is installed with engine.

Also changed not to install the auth provider separately from the engine

Tested on both 4.0 and master.
On 3.6 it just doesn't break anything, but doesn't tested the log collector.

Change-Id: Ie96c7c7d7ff8ba0bb06260d0fe8a71cde36da8c2
Signed-off-by: Yaniv Kaul <ykaul@redhat.com>