vdsm_recovery test
When Vdsm is stopped for any reason (e.g. by admin or due to a crash),
statuses of its VMs should change to "unknown" on Engine soon.  Once
Vdsm is started again, correct statuses should appear in Engine sooner
or later.  With this patch we introduce a simple initial test for that
functionality.

Change-Id: I9902df9f240f70ee4caf0e128e9c1ca439d26782
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>