Switch to use virtio-scsi
Now that Lago supports virtio-scsi, it's time to use it.
Switched master, 4.0 and hosted-engine to use it.
It provides slightly better performance, and with 'fstrim' being used,
less disk space is being consumed for running tests.

Tested on master, 4.0, hosted-engine and 3.6.

It also required a small change in HE resources.
It can run with less memory well, so I reduced it.
This way, it succeeds on a 8GB RAM host.

Change-Id: I9b216f9ac58d591842b29191b8548bd19d8037d4
Signed-off-by: Yaniv Kaul <ykaul@redhat.com>