Allow a custom reposync config in run_suite.sh
This patch adds a --reposync-config option to run_suite.sh,
if not passed it will use the default.

This option can reduce the number of suites we need, as sometimes(
not always) the only difference between suites is the repository
links. It will allow adding a 'reposync-4.0-stable.config' for instance,
instead of creating a new redundant suite just for the stable
repository.

Change-Id: I62352d82b582fb5b67bb18977134cf6214d94960
Signed-off-by: Nadav Goldin <ngoldin@redhat.com>