CHAP and IPv6 for iSCSI
This patch configures the iSCSI target to support both IPv6 and CHAP auth.
I could not make yet CHAP discovery mode on oVirt, so only login with CHAP
(initiator) is enabled for the time being.
IPv6 is configured but not used yet.

[Patch 2] Additionly in the patch, we now only enable multipath, login to the iSCSI target,
take the GUIDs, disable multipath it on the host and logout from the target.
Previously we used to log in and keep multipath running.

[Patch 3] Separated between loading multipath and logging in to the iSCSI target and reading
the multipath output, to allow it to actually get all the data.

[Patch 4] Make the change in 3.6 as well

[Patch 5] Restores the HE hosts memory to previous values, delete iSCSI target data.

[Patch 6] Find out WWN in a different way, drop multipath altogether
Change-Id: Id2cf99210a3bc60a020c151de01f4c2a7d107429
Signed-off-by: Yaniv Kaul <ykaul@redhat.com>