Added a feedback on failure to gerrit trigger
We can split the tests into two parts.
part 1 - creating the env.
part 2 - running 00*.py tests.

In case of failure in the test, create a specific
message in 'failure_msg.txt' which will be shown
in gerrit.

This will ease the process of understanding the cause
which made a test to fail.

Change-Id: I78996d03fc9c742fccdc603dbe4c918458d2a68e
Signed-off-by: gbenhaim <galbh2@gmail.com>