test scenarios: reintroducing migration tests
Commit f0ba28da4a removed the migration tests because we sometimes
failed to set an IP address on the VLAN200 network,
getting a "bridge already exists" error in supervdsm.log.

Not setting the VM role (only setting migration role) avoids the issue
as a bridge is not being created. In fact, the VM role was redundant
to begin with.

In order to use the non-VM VLAN, MTU had to be set to match the other
networks on the NIC to the default value. An attempt to use the
recently added bond in 003 was made, however it wasn't successful;
although a qemu connection was seem to be made, the migration process
timed out with no migration taking place.

As it's important to have migration tests present, I'll investigate
the bond migration failure at a later time and resolve to modify
VLAN200's MTU instead.

Ipv6 migration is currently commented out due to missing ipv6 iptable
rules (bz#1414524).

Change-Id: I59661425b37164b37018728d1a526161ca155cdc
Signed-off-by: Leon Goldberg <lgoldber@redhat.com>