Disable http proxy in run_suite
Since reposync is configured to work with mirrors and
repoman downloads rpms from resources.ovirt.org, there is
no need to use a proxy.

The proxy will be only disabled during "lago ovirt reposetup",
which means it's going to effect only on reposync and repoman.

This will help to reduce infra issues in the tests.

Change-Id: I20bb3398a69e97b38ed3f7e72d257bc1016d1e96
Signed-off-by: gbenhaim <galbh2@gmail.com>