Fix the non responsive host issue when using add host w/ SDK v4
The 'override_iptables=True' is missing from host configuration,
which causes the iptables to be configured with only the SSH port
being available by default, which prevents the ovirt-engine from
connecting to the host's VDSM.

The option was set by default in the test code that used v3 SDK,
but got missing from the add_hosts code that uses v4 SDK, thus
the 'non-responsive host' issue.

This patch fixes the issue.

Change-Id: I4695c1860f557d0d5d9b22bc7812db0f1bfdf673
Signed-off-by: Lev Veyde <lveyde@redhat.com>