ost: added fallback reposync after the initialization of the base version
I've encountered a bug where upgrade job fails on missing rpms when upgrading
from previous release (upgrade-from-prevrelease-*)
This is because 'latest' is not a full repo and have only part of the rpms.
So sometimes, when upgrading the engine, some rpms which might be mandatory for
the upgrade wont be in the 'latest' repo.

Change-Id: Ie62790c3d4c56ab27da9547789d3d9fdd36525e6
Signed-off-by: Daniel Belenky <dbelenky@redhat.com>