Add a second NFS storage domain.
For now, the extra storage domain will be useful in storage
domain reattachment tests, which follow.

Later, it will be switched to using IPv6 and will be set up
over the storage network (as soon as Lago learns to resolve
hosts over these, and DHCPv6 is used on all nodes).

As for common/deploy-scripts/setup_storage_unified_el7.sh,
I have seen "Deploy oVirt environment" take 3:13 or 3:36
on master (but I didn't run it many times). With this patch,
it takes around 3:13, or 4:43.

In basic-suite-master/test-scenarios/002_bootstrap.py,
add_secondary_storage_domains typically takes either 2:28
or 2:38 on master. With this patch, the times were 2:26,
2:34, and 3:04 (twice). So sometimes, this test might add
half a minute to the runtime.

On a Jenkins slave, the test took 0:02:30 for this patch
as well as the second one, and 0:01:52 for the 3rd patch,
so we cannot really talk about penalty if there is such
variance between test runs.

Change-Id: I849c36468e9f8f8070603e42716e41a55fe475da
Signed-off-by: Ondřej Svoboda <osvoboda@redhat.com>