Add a VM with a custom MAC to the second SD.
The VM will disappear once the domain is detached in tests
which follow. It will be imported back with its MAC reassigned
to be inside the MAC pool.

In total, this patch may add up to 9 seconds to the runtime.

add_vm_blank usually takes 0 s on master and between 1 s to 2 s
with this patch. add_nic always takes 0 s. add_disk takes 10 s
on master and between 8 s to 17 s with this patch.

Change-Id: I05bcc7ce7dc53d1f257fbd5791dcbcbfe2a7b5d0
Signed-off-by: Ondřej Svoboda <osvoboda@redhat.com>