ost: removed extra_sources from check-patch
We used extra_sources when we had nightly rpms, and we
had to make sure that suites wont fail due to 'bad rpms'.
As we are not using nightly and latest_tested repo, we no longer
need this change.

Also dropped the 'exclude' list from tested repo, as we need all the 
repo usually.

Change-Id: I96566c9be040b49abcaae8b8f02611d4aaa23962
Signed-off-by: Daniel Belenky <dbelenky@redhat.com>