Rename VLAN200_Network to Migration_Net
This is its current usage, only not shown in the name.
We cannot call it Migration_Network due to interface
naming limitation (up to 15 characters).

The network is attached to hosts in 003_basic_networking
for the first time (and immediately detached). In 006_migrations,
it is attached again.

If we move networks teardown to a later time, we can
save one network setup.

Change-Id: I984bc73e75b9b112a9253b9d16ba8d1f63a901c4
Signed-off-by: Ondřej Svoboda <osvoboda@redhat.com>