4.1 backport: move migration tests to 006_migrations
This patch is the last one to preserve network tests
which use SDKv3 in 4.1 before we migrate them to
SDKv4 in master.

Since it is not possible to configure an IPv6-only
(migration) network in 4.1, we only backport an IPv4
migration test.

We will not rebase storage domain reattachment tests
because they add a minute to runtime and only use SDKv4.

Change-Id: Iaf23f9a9742e8907fd946377088950a78a5fd955
Signed-off-by: Ondřej Svoboda <osvoboda@redhat.com>