Changing 004_basic_sanity.py::snapshot_live_merge
This patch undo the 'snapshot_live_merge' test's skipping by removing
the 'raise SkipTest' line.

This patch also enforces the snapshots to be taken for the image disk.
Without this enforcement, using 'disks.list()[0]' doesn't guarantee
that the first disk is an image disk and not a dlun disk (which is also
been created on this test).

Change-Id: I25852aab4070fdd3eb9459174990e6cf66ddbffb
Signed-off-by: Shani Leviim <sleviim@redhat.com>