Adding scenario testing ovirt-provider-ovn
This patch checks if ovirt-provider-ovn
responds to request after being installed
by engine setup.
The tests makes an authorization request, checking
if the recieved token is of correct format, and
a basic get networks request (receiving an empty
list of networks).
Note, that this patch does not check the correctness
of the certificates used for the SSL connection.
This will hopefully come in later patches.

Change-Id: I0ee1ff01da43ed72544d766aa89e07dc5fcf8ace
Signed-off-by: Marcin Mirecki<mmirecki@redhat.com>