Introducing performance suite
This suite will help to deploy OST tests on large environments. It has
2 main features:
1. Deploy hosts and VM using fakevdsm, or use real(default)
2. Set the number of hosts and number of VMs you want to deploy

When this suite is proven stable and useful, its features can be ported
to the 'basic-suite' and itself can be removed. So basic-suite-master
for example can scale without problem.

Motivation:
Deploy OST using variable setups, gather stats and analyze it.

Change:
- Env variable for number of Hosts
  OST_HOST_COUNT
- Env variable for number of VMs
  OST_VM_COUNT
- Env varibable to use vdsm fake (default unset)
  OST_USE_VDSMFAKE
- A stripped down test sceneraio which uses a pool of thin, headless
VMs, prestarted

Change-Id: I4d07fa8feb3d8ae08125d44f35841e410444c26b
Signed-off-by: Roy Golan <rgolan@redhat.com>