Adding change resolver to check-patch for better changes detection.
Jira: https://ovirt-jira.atlassian.net/browse/OST-62

As in the past weeks, we've had few undetected errors in OST,
and sometimes we've run irrelevant suites on some of the commits,
 we've decided to improve the logic of check-patch.
The detection logic now holds a cached map (Will be generated only once per
instance, and only by demand) which maps every file to the suites it relates
to. This allows us to run only relevant suites on a given change
(file from the commit which is being checked).

The logic is split into 4 base cases:
1. A change was made to a core file of the runtime scripts
   (check-patch, suite, change_resolver):
   In this case, we'll run all the suites we've detected in the root directory
   of OST.
2. A change was made to an automation script or config file:
   In this case, we'll add the related suite, and search for symlinks that were
   pointing to the changed file. We'll also check the suites related to those
   symlinks.
3. A change was made inside a suite's directory:
   In this case, we'll check the related suite, and search for symlinks that
   were pointing to the changed file. We'll also check the suites related to
   those symlinks.
4. A change was made to a common file.
   In this case, we'll search for all the symlinks that were pointing to this
   file, and run the related suites.

Change-Id: Ifd7fd882d539f0bba3a89cfa1da98cb318d69476
Signed-off-by: Daniel Belenky <dbelenky@redhat.com>