Fix storage-domain 'active' check
We used to check if the SD was active only if it was in maintenance
first. What if it was in a different state?
Now we check that it is in active state regardless of the previous
state. Before that, if it's in maintenance, we activate it, no change
there.

Change-Id: I635ea685564433a6771f90a418c83bee3c0aabf5