Add metrics to basic suite
Motivation:
Collect metrics on a suite execution and use it for to analyze and
characterise the system behaviour in different flows.

Change:
Use ovirt-engine-metric to deploy collectd and fluentd on every active
host and the engine. And make the engine the fluentd aggregator

Result:
Metrics are collected during the execution and are written to a
file under /var/log/fluentd/fluentdout-${hourly-timestamp}

Change-Id: I3edf0561f1c440d05bea89b535c68d83bb35cc8d
Signed-off-by: Roy Golan <rgolan@redhat.com>