Add a 1 second sleep after LSM
This patch is a temporary hack to fix a race condition[1].
The patch[2] sent to initially resolve the issue introduces
other, more severe, problems[3] and will be reverted.

A better solution will be used once the RFE[4] is
added and will provide a sensible solution to polling
the job's status.

1 - https://bugzilla.redhat.com/1456504
2 - https://gerrit.ovirt.org/#/c/77504/
3 - https://bugzilla.redhat.com/1460694
4 - https://bugzilla.redhat.com/1460701

Change-Id: Idced0d072d32918bbb8b41c9c66953cd6d9de3b3
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>