Ping vm0
After vm0 is booted, it is pinged. This checks multiple things, e.g.
it ensures, that the VM is created, is able to boot the OS,
has a connected network interface, network filtering does not block
the communication and vm0's network interface has the expected IP
address.

This test takes usually less than a second.

Change-Id: Ic56efeb29542e53ce6d4c9293d8d4af5f9c6fe96
Signed-off-by: Dominik Holler <dholler@redhat.com>