ovirt-provider-ovn: Test adding/removing entities
The test verifies that ovirt-provider-ovn was installed
and configured propertly, and that some basic rest
operations can be performed on it.
In the test we add one instance of each Neutron entity
supported by the provider (network, port, subnet),
verify that they have been added, delete them, and
verify that they have been removed.

Signed-off-by: Marcin Mirecki <mmirecki@redhat.com>
Change-Id: I51121ed867dc9cfa329cc38727d127f1f42e2ed6