ovirt-provider-ovn: test import/remove external network
This patch tests the import of an external network to
ovirt engine, and then it's removal.

When provider host deploy is ready, the test will have to
be split in two. The networking setup will have to happen
before hosts are set up, tests checking the external
networks on hosts will have to be added, and the tests
which remove the external networks will have to be done after
the vm is removed.

Test time < 30seconds

Change-Id: I8832f5cfc9018e31514f481a2b8946fe5db0dc9a
Signed-off-by: Marcin Mirecki <mmirecki@redhat.com>