skip repo closure test for master suite
During the last weeks this test has been failing multiple
times, sometimes due to real issues, mostly in 'ops tools testing'
repo, but many times due to caching problems or other errors
that are infra related.

Bottom line, it is causing us more noise and work then
the value it brings, so until we can find a better solution
we will disable it on master to regain stablity.

A few options to consider:
1. move it to its own suite and run nightly
2. improve code so it will only check for missing pkg in oVirt repos
3. get rid of unstable repos that might cause repoclosure to fail
   more often

We will keep monitoring master suite for any new errors that
might pop, if we'll see that the removal of the test is causing
errors on oVirt tests that are harder to debug, we can consider
re-enabling the test, hopefully with option '2'.

Jira: https://ovirt-jira.atlassian.net/browse/OST-71

Change-Id: I551396ad561d1bd1cc7c9f566d034b2974f9443d
Signed-off-by: Eyal Edri <eedri@redhat.com>