Change add hosts flow
1. Do not install all hosts in parallel.
It probably isn't worth the effort, does not really reduce runtime.
It is probably not a great idea while installing many hosts anyway
(performance suite, for example)
2. Verify that hosts were installed separately from the installation
itself. In between we can run few other tests (with the hope we can
add even more in the future).

Overall, there is no real change in the time it takes to install a
host. If anything, it seems to be a bit lower on my laptop, but not
significant. The code is somewhat cleaner, I hope.

Change-Id: I1a5bcbf12e5ee9d636f8fc33abd264df08f7c682