Install ovirt-host on the 1st host only.
We cheat a bit, and during ovirt-engine installation on Engine
VM, we also install, only on the 1st host, ovirt-host meta package.

This allows to get the 1st host up and running faster, which allows
getting the master storage domain faster, which allows getting the
secondary storage domains and Glance imports faster, and everyone's
happy.

We are still checking real host deployment on host 1 and on (if there
are any, for example in the performance suite).

It adds ~50 seconds to the initial deployment, but should pay itself
off later.

Change-Id: Iefe4a0e979ef2475b4510f673ae769e6ecb5ff8a