Introducing networking suite.
With OST continously expanding, both performance and maintainability
are becoming pressing issues.

We have recently thought of introducing testing scenarios for mac pools,
but that proved to be too costly; the scenarios potentially require a
lot of VM operations that take a lot of time. These scenarios, although
not entirely network specific, are looking to test code that is
networking based.

As a result, the idea for a separate OST suite that is dedicated for
networking is being brought to fruition. The suite is pytest based
which will ease handling and creating test objects via dedicated
fixtures.

Unlike current OST suites, tests are not going to be co-dependent,
with the aforementioned fixtures taking care of object creatinon/
destruction without having to use excessive boilerplate code.

Additionally, tests will be able to be ran individually. This will
hopefully greatly reduce the time spent to verify specific changes.

Change-Id: I6dbb9a7c73cc52fa900ed50b184145ac4e33f136
Signed-off-by: Leon Goldberg <lgoldber@redhat.com>