Add OVN Provider to cluster
The oVirt Provider for OVN is used as default provider for
external networking in the 'test-cluster'. This way the
hosts are enabled to use OVN automatically.
This prepares testing OVN networking in OST.
This change adds around 10 seconds execution time, because
the installation of the ovirt-provider-ovn-driver and its
dependencies is checked (and performed if required) during
adding the hosts.

Change-Id: I34629cdee9fef45a989b484f858db34f8fdaf330
Signed-off-by: Dominik Holler <Dominik.Holler@gmx.net>