Change function name
Change the name of 'test_ovn_provider_rest' to 'use_ovn_provider',
because the old name starts with 'test_' and the function was
unintended detected as an additional entry point to the file.
This way the 'test_ovn_provider_rest' was executed twice.

This change saves around one second execution time.

Change-Id: I035a9ee8294445004ca15327533b7c4ff0696f0d
Signed-off-by: Dominik Holler <Dominik.Holler@gmx.net>