Begin centralization of constants
Currently every python file manages it's constants
internally. To improve the maintainability, the constants
are started to centralize across all files of a test suite.

The first constant which is centralized is the
name of the default ovn provider.
Because the id if this provider is dynamic, it is
determined by a getter function in network_utils_v4.

Change-Id: I472a15d8100cc7e1b90c83d1cd97570aa5fb1b8e
Signed-off-by: Dominik Holler <Dominik.Holler@gmx.net>