reload postgres instead of restart
After tweaking the pg configuration, just reload it instead of a restart
otherwise tests might fail because the engine wasn't fast enough to
renew the connection to the db.

As long as the configuration tweak is not something which demands a
restart this is good.

If in the future we want to tweak a config item that demands a restart
to take effect we must wait in the script till engine renews the
connection:

```bash
 test 200 == $(curl -k https://ovirt/services/health  -o /dev/null -s -w %{http_code})
```

Signed-off-by: Roy Golan <rgolan@redhat.com>

Change-Id: Ib659f4b7194b8245552c393473e6ee73d6348553