Reboot vm0 after resume
VM networking may not work after resume.  It seems to be a guest OS problem so
the best solution would be to use another guest OS, such as CentOS or Fedora.

Before we contingently switch to another guest OS, let's use a workaround to
unblock tests that require access to the VM's guest OS and are prevented from
that due to non-functioning networking.  We reboot the guest OS, which fixes
the problem.  We use virsh rather than Engine/Vdsm to initiate the reboot in
order to prevent restart of the whole VM, not to hide contingent bugs related
to suspend-resume.

Change-Id: I8c89e3d7945666d96f8e21f5a89269e14165e073
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>