fluentd aggregator floods the disk, make it optional
Motivation:
fluent out plugin is about to be deprecated. At the moment the gzip
capabilities seems not to work, so the disk is flooded with gigs of info
and the vm stops working.

Modification:
Since this functionality is not critic to have
as different methods of monitoring exist, I'm switching it off by
default.

Result:
fluentd will not write its events to a file on the disk.

export OST_FLUENTD_AGGREGATOR=1 if you want to switch it on.

Change-Id: I34da5492d5b32a0b3e3b078886eff961d3709122
Signed-off-by: Roy Golan <rgolan@redhat.com>