No NetworkManager restart FirewallD installation
The restart of NetworkManager service after FirewallD installation
was introduced in commit 58c9d243f2ae6cc74ff8c75d13de76a556b009d3
as a workaround for BZ 1436770. Because this bug is fixed now, the
restart is not required anymore. As a nice side effect, the
deploy-scripts does not depend on NetworkManager anymore.

Change-Id: Iab96fcf60548dc418f69bb8baab5aa5e6477b2c7
Signed-off-by: Dominik Holler <dholler@redhat.com>