network-suite: define migration network as non-vm
Before this patch, mig-net is attached to each host with a bridge since
it is defined as a VM network on the data center. VM usage is an
unfortunate historical default

This patch lets users of netlib.Network control whether it is a VM
network by setting usages to either [] or [vm]. It also defines the
migration network as a non-VM one.

Change-Id: I03db1351e8dae0ad0dec65ad3683f1bea87cd146
Signed-off-by: Dan Kenigsberg <danken@redhat.com>