4.2: add another official ovirt-4.2 repo as fallback to tested
As part of the mirror process [1], we inject the oVirt tested
repo instead of $ovirt_release reponame, which we usually use
for official releases.
This usually works ( 4.1 currently ) since tested already
include all of the RPMs.

However, in an early major release, like 4.2 is now, not all pkgs
are yet in tested since not all projects branched yet to 4.2.
This leads to a status where the reposync file has 2 repos
pointing to tested and thus missing a lot of RPMs.

This patch adds the official 4.2 repo as fallback so any missing pkg
from tested will be taken from the official release

[1] http://ovirt-infra-docs.readthedocs.io/en/latest/CI/Transactional_mirrors/index.html


Change-Id: I0103d5e7c835fd6127b8d6539635b0f20bf9c468
Signed-off-by: Eyal Edri <eedri@redhat.com>