Use ansible to access ovirt-provider-ovn
The basic idea of this test is to utilize an existing
client implementation of the OpenStack Networking API to
access ovirt-provider-ovn's server implementation of
OpenStack Networking API to check API conformance.

Ansible is very suitable as an existing client to test this,
because ansible does not just create requests, but also
checks if the requested changes are reported as expected, if
the same playbook is executed a second time.

The added test uses ansible to create OVN entities on the
ovirt-provider-ovn, ensures that they are reported, removes
them and ensures that they are removed.

This initial commit creates just a minimalistic scenario.
Complexity can be added to the scenarios just by adding
entities to the ansible playbooks.

The playbooks can be executed locally by
ansible-playbook network-suite-master/ansible/create_scenario.yml
after creating the file
network-suite-master/ansible/clouds.yml with content like:

clouds:
  ovirt:
    auth:
      auth_url: https://0.0.0.0:35357/v2.0
      password: '123456'
      username: admin@internal
    verify: false

Change-Id: I9d1273d7a913206f4c6e2f087c49996a303e5546
Signed-off-by: Dominik Holler <dholler@redhat.com>