networking: Introducing mac pool tests
This patch introduces a module for mac pools tests and the relevant
infrastructure.

Two tests were added; a test that checks that using an address that is
present in a stateless snapshot in a different vm fails, and another
one that checks the aforementioned scenario succeeds for a stateful vm.

The relevant infrastructure consists of Vnic related code that allows
the creation and manipulation of Vnics.

In the future, additional testing scenarios should be added, such as
checking whether an address present in a non-stateless snapshot is also
unavailable for other VMs.

Change-Id: I1ae68e5712fc9049e3ac9e56a8aa33226fe3fbad
Signed-off-by: Leon Goldberg <lgoldber@redhat.com>