upgrade_suites: Update ovirt-engine-metrics
Since cd7c467dde53c57243d3d933ea4be7b6d6dcf1b5 in ovirt-engine,
'ovirt-engine-metrics' needs to be installed in order to add host.

Currently, the upgrade suites run 'yum -y update ovirt-*setup*', which
doesn't update 'ovirt-engine-metrics', thus 'oVirt.metrics' role is
missing and add host fails.

This commit adds an update for 'ovirt-engine-metrics'.

I'm not sure if this is a bug in the test, or in 'engine-setup',
but for now adding this fix in order to not break CI.

Change-Id: Icda62d6b4884a9bda6cfa7f5877fb8d49b517f08
Signed-off-by: gbenhaim <galbh2@gmail.com>