Test snapshot with memory
Let's introduce a test restoring from a snapshot with memory.  Note that we'd
like to test not only that the restore operation works, but also that following
common operations work too (at minimum migrations).  On the other hand some
operations, such as disk hot plug, are not possible while previewing a
snapshot.  So placement of the test matters from the view of tested
functionality.

This is just a basic version of the test to be extended in followup patches.
Future versions of the test should check for instance that the right drive is
used or (in a different suite) that restore from snapshot with memory works
across cluster upgrades.

Change-Id: Ibee6681ccfac191ab6a5fe6564bdde055f3623f6
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>