Add a check for resume log items to suspend_resume
The current suspend_resume test checks that a VM is up after resume.  But it
doesn't check how the VM got up.  If the VM was simply restarted rather than
resumed, the test passes.

This patch adds a check for traces of resume operations in Vdsm log.  That's
not a very elegant method but it's used due to lack of better options.

The enhanced test still doesn't check whether the guest OS survives the
suspend-resume operation.  That's checked in a followup patch that is probably
not going to get working with Cirros guest OS.  Once the followup patch is
working, it will obsolete this patch.  Let's have this patch until then.

Change-Id: Iec942fcffb99a61145a8fc8bee3ef0eb423bf84b
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>