networking: collect artifacts even with engine failure
In case engine fails to setup, we still want to collect
artifacts. Currently, in case engine's fixture fails, artifact
colletion fixture is not reached.

The naive solution of requiring artifact collection fixture in
engine's fixture does not work as engine's scope is greater than
artifact collection (session > module).

In order to collect artifacts even in case of a failure,
implmenet artifact collection logic in testlib and use it in
engine's fixture.


Change-Id: Ie101bfa4e692507c9e8458ed264feb19c482dda1
Signed-off-by: Leon Goldberg <lgoldber@redhat.com>