Make sure that ssh is working after resume
Networking may be broken in the vm0 guest OS after suspend-resume.  We check
for that situation in restore_vm0_networking and reboot the guest OS if
networking is not available.

However it may also happen that ping is working but ssh timeouts.  That has
been observed in hotplug_cpu test.  To prevent that situation, this patch adds
also ssh check to restore_vm0_networking.  If it doesn't pass the VM guest OS
is rebooted the same way as when networking is not working at all.

Change-Id: I4bbff585281bb29fbe82faf20df410b1d105cf20
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>