he-iscsi-master: Temporarily exclude in check-patch
When running in the CI, we check if there are 15GB of free
space in /dev/shm, if so we create the Lago env there.

he-iscsi-master requires ~70GB of RAM, thus it fails check-patch.

Until we add the functionality that allows each suite to specify its
requirements, exclude he-iscsi-master in check-patch.

Change-Id: Icc29bbff5cb7bc312b495a25bf7be934d83e7238
Signed-off-by: gbenhaim <galbh2@gmail.com>