Test for events
Added, for some but not all tests, some simplistic search for
a matching event, by its ID.

It's not complete (we can add more), but it's a good start.

It was created in a response to a silence failure in a command
that actually had a failure event, but it was not noticed.

Change-Id: Ia2c9b1dcec66fd5b06bc7404647815c2dd936d9c