network: Avoid python cache usage when running the tests
When running two suites or more on the same tests, the python cache
files (*.pyc) are not refreshed, although the context from which they
have been triggered has changed.
Specifically, when running  network-master suite and immediately after
the network-4.2 suite, the later is failing on the session fixtures.

We do not really understand why the context matters.
It might be related to special handling by pytest.
At this time, we prefer to avoid *.pyc instead of keeping to research
the mode of failure.

Change-Id: I1f950fc9e89aa9732464a95d684c4cd9ef28f781
Signed-off-by: Edward Haas <edwardh@redhat.com>