Enable auto_sync after import
During the network is created and the network is imported
in test 'use_ovn_provider' auto_sync has to be disabled to
prevent the manually created external network on the provider
from being imported automatically. After the import,
auto_sync is reset to its original value, as this is what we
recommend our users to do.

Change-Id: I5a680f907feb33bb6249a0b1fea5dd77723b7422
Signed-off-by: Dominik Holler <dholler@redhat.com>