Enable Vdsm debug logs
When hunting problems revealed by OST, it's very useful to have debug logs of
certain Vdsm loggers.  We enable them using vdsm-client to avoid the need of
Vdsm restart.  But we also set them in logger.conf so that they remain
applicable after Vdsm restart, such as in the recovery test.

Change-Id: I8dc63fd2a646f38f9c165d46501b61693f71bc45
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>