network: Statically set pip and shade versions
pip and shade packages are installed/upgraded using the pip tool.
It is safer and more stable to install the packages in a controlled
manner by fixing their version.

Recent pip updates, introduced pip version 10, which deprecated the
ability to uninstall distutils sub-packages, specifically, failing
the uninstall of PyYAML package which is required by shade.

Change-Id: I3970c3857caaf84aa157a9f92baff6018a4c3146
Signed-off-by: Edward Haas <edwardh@redhat.com>