Allow to skip collection of logs on successful scenario execution.
If the scenario succeeded, no need to collect the logs. Useful mainly
when developing tests, but also generally, good idea.

Also had to change a bit the test for downloading of files from Engine:
the collection slowed the process and allowed the Engine to be up to collect.
However, without collection, you really need to retry.

Change-Id: Ic0fbd413dd5112841d48dbab723764e6e49c21f6