network: add test validating connections between OVN  networks
This patch test the connection between OVN networks.
It adds (and deletes afterwards) internal ovs network
ports, and binds them to OVN logical ports.
These ovs ports are then used to test connectivity
between the ovn ports.

Note that this is a temporary solution until we are
able to ping via the deployed vms.

Change-Id: I4ca08e609d67e920664e6ef2ce5a31b47c7e1f4d
Signed-off-by: Marcin Mirecki <mmirecki@redhat.com>