network: Host.remove_networks() by Network object
To remove network attachments from a host, it is not possible
to specify a network name because the network sdk type does not
carry the name of the network and so the name cannot be used
to locate the required network.
Instead, the netlib.Network objects will be passed and their ids
used to resolve the networks.

No tests failed due to this issue because until now this
functionality has only been used in tests teardown.

Change-Id: I1018b1cc5e9bdac370ca35c9ba3c381e4b6e91f4
Signed-off-by: Eitan Raviv <eraviv@redhat.com>