network: Modify ovn topology created by ovn to test external_gateway
A second router with another network behind it is added
to the network topology created by the ansible 'create-ovn-entities'
role.
This is needed to test if packets addressed to a network which is
not directly connected to a router will be routed correctly when
the external_gateway_info is specified.

The network topology created in this patch looks as follows:

           [net14]     --▶ net14_port1 10.14.0.2
           (net14_subnet1 - 10.14.0.0/24)
             |
             | 10.14.0.1 (gw)
             |
             ▼
           router1
             ▲
             |
             | 10.12.0.1  (gw)
             |
            [net12]    --▶ net12_port1 10.12.0.2
            (net12_subnet1 - 10.12.0.0/24)
             |
             | 10.12.0.100
             |
             ▼
           router0
             /      \
10.10.0.1   /        \  10.11.0.1
  (gw)     /          \    (gw)
          ▼            ▼
      [net10 ]        [net11]
  (net10_subnet1     (net11_subnet1
   10.10.0.0/24)      10.11.0.0/24)
       |                 |
       |                 |
       |                 |
       ▼                 ▼
    net10_port1        net11_port1
    10.10.0.1            10.11.0.1

Router0 is using net12 as external_gateway_info network.
This is an equivalent of adding a default route via 10.12.0.1
on router0.

The net12 represents the OpenStack external network, which is
usually created by the administrator. In long-term net12 will
have a physnet and the routers connected to net12 will
provide SNAT. Neutron requires that the attribute
router:external of net12 is true.
If the scenario is applied to an environment outside this
test, the removal of the public network has to be skipped.

Change-Id: I5e300d3c0ce7d487a518eb1add06438e777a197c
Signed-off-by: Marcin Mirecki <mmirecki@redhat.com>