repos: adding cockpit-machines-ovirt
Adding cockpit-machines-ovirt.

Note that this is available in CentOS Virt SIG only,
not being shipped in extras.
I see that hosted engine repos explicitly exclude Virt SIG and
uses extras forcefully even if it has a lower version.
Not knowing the reason behind that choice I'm not going to touch
those repos with this patch.

Change-Id: I72ead603adc244adaae21dff57cf48f530d6e5c3
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>