reflect package python-netaddr rename to python2-netaddr
https://gerrit.ovirt.org/#/c/91775/ introduces the renaming of at
least the package python-netaddr to python2-netaddr.
It's a proper renaming following python packaging guidelines, because
python2-netaddr obsoletes and provdes python-netaddr, see
http://cbs.centos.org/koji/rpminfo?rpmID=124829 .

ovirt-provider-ovn depends on ovsdbapp, which depends on
python-netaddr >= 0.7.18 , but the version dependency is not yet
reflected in the packaging, see
https://bugzilla.redhat.com/show_bug.cgi?id=1553268
This is no problem, because yum will pick a recent version of
python-netaddr, if it is avaliable in the repo.
Since https://gerrit.ovirt.org/#/c/91775/ renamed python-netaddr to
python2-netaddr, now a recent version of python2-netaddr is required in
the repo.
Even if this fix makes all related suites in OST succeed again, this
might be inclomplete, because all packages in
common/yum-repos/ovirt-master.repo have to be checked if they are
renamed in https://gerrit.ovirt.org/#/c/91775/. If this is not checked,
OST might use other packages than available in CentOS Virt SIG oVirt
4.3 repos.


Change-Id: I91c6c936a64844726a61f2e331b7823f40d3a96b
Signed-off-by: Dominik Holler <dholler@redhat.com>