network: introduce wait for 'powering up' or 'up'
In some tests, 'powering up' state is enough to indicate the test
has passed saving a wait for 'up' status. This can save execution
time for the suite.

Since 'powering up' is a transient status and waiting on it might
cause might be too late, this facility waits for either 'powering
up' or for 'up' status.

Change-Id: Icff4449adcac1607cc02698a444f00e10efe6433
Signed-off-by: Eitan Raviv <eraviv@redhat.com>