network: Add test for provider physical network
A VM connected to a provider physical network is created and started.
The tests adds more than 80 seconds of execution time, because it
waits for the VM to be in the state UP.
In follow up commits will be improved to reduce the waiting time.

To remove an external logical network, the network has to be removed
directly on its provider by OpenStack Networking API.
The entity representing the external network inside oVirt engine
has to be removed explicitly here, because auto_sync is disabled for
the provider.

Change-Id: If33c802190ff61fbf1b9afae860419fe87e7d28f
Signed-off-by: Dominik Holler <dholler@redhat.com>