network: Propagate the network MTU to the VM
Update the created physnet to have an MTU of 1000 bytes.

The ping command had to be updated, now having the packet size
as a parameter.

The test itself attempts to ping a VM from the host that is not
running it with a packet size lower than the MTU (valid ping)
and afterwards attempts a ping with a packet size larger than the
MTU, which results in a thrown exception.

Change-Id: I07b6661b9f06d3004fa5adb79736078672bc02d3
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>