network: Workaround for missing shade MTU functionality
This is a workaround for missing shade functionality
Because of functionality gaps in shade we are using internal shade
_network_client to manually contruct the required query.
The alternative would be to handle openstack authentication, path lookup,
etc. manually. Here the cost is accessing _network_client
The shade bug to add the missing functionality is BZ1590251

The patch also adds a test that modifies the mtu of an existing
ovn network.

Change-Id: I2800e73340748185573356260027b950b8c10769
Signed-off-by: Marcin Mirecki <mmirecki@redhat.com>